chore(api): remove legacy monitoring card-data UI route
Closes #599. The dashboard and monitorings index now fetch card/summary
data through the versioned api.v1.internal.ui.monitorings.cards endpoint
(commits a95657d/985ae43); the legacy /api/monitorings/card-data route
and MonitoringCardDataController had no remaining callers besides their
own tests. Regression coverage (query-budget, batching, per-user
scoping) is ported onto the surviving endpoint instead of dropped.

Also fixes a stale AGENTS.md reference: the webguard-instance
compatibility-sensitive routes live in routes/api/instance.php, not
routes/api/internal.php, since the #593 API-boundaries refactor split
scanner and browser-UI internal routes.
